:truck: Rename and move jwt-payload.dto to interfaces/jwt-payload.interface
