+++ backend/src/users/interfaces/jwt-payload.interface.ts
@@ -1,4 +1,4 @@
-export class JwtPayloadDto {
+export class JwtPayloadInterface {
   email: string;
   permissions: string[];
   id: number;
