+++ backend/src/users/strategies/jwt.strategy.ts
@@ -1,8 +1,8 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, ExtractJwt } from 'passport-jwt';
-import { JwtPayloadDto } from './jwt-payload.dto';
-import { User } from './entities/user.entity';
+import { JwtPayloadInterface } from '../interfaces/jwt-payload.interface';
+import { User } from '../entities/user.entity';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
@@ -13,7 +13,7 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
     });
   }
 
-  async validate(payload: JwtPayloadDto) {
+  async validate(payload: JwtPayloadInterface) {
     const { email } = payload;
     const user = await User.findOne({ email });
     console.log(user);
